Starting to move common code into bag2video_common
