+++ .gitignore
@@ -34,3 +34,4 @@ nosetests.xml
 .project
 .pydevproject
 .venv/
+.pdm-python
