+++ bag2video_rosbags/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["get_sizes", "get_frequency"]
+
+from bag2video_rosbags.utils import get_sizes, get_frequency
